Update Task to use list of statuses & create LazyListScope extension to display task cards

+++ core/data/src/main/java/com/nrr/data/repository/RoomTaskRepository.kt
@@ -6,7 +6,6 @@ import com.nrr.database.model.ActiveTask
 import com.nrr.database.model.TaskWithStatus
 import com.nrr.database.model.asEntity
 import com.nrr.database.model.asExternalModel
-import com.nrr.model.ActiveStatus
 import com.nrr.model.Task
 import com.nrr.model.TaskPeriod
 import kotlinx.coroutines.flow.Flow
@@ -42,28 +41,27 @@ internal class RoomTaskRepository @Inject constructor(
             it.map(ActiveTask::asExternalModel)
         }
 
-    override suspend fun saveTasks(
-        tasks: List<Task>,
-        activeStatus: List<ActiveStatus?>
-    ): List<Long> {
+    override suspend fun saveTasks(tasks: List<Task>): List<Long> {
         val ids = taskDao.insertTasks(tasks.map(Task::asEntity))
-        activeTaskDao.insertActiveTasks(
-            tasks
-                .filter { it.activeStatus != null }
-                .map { it.activeStatus!!.asEntity(it.id) }
-        )
+        tasks
+            .filter { it.activeStatuses.isNotEmpty() }
+            .flatMap { it.activeStatuses.map { s -> s.asEntity(it.id) } }
+            .run {
+                if (isNotEmpty()) activeTaskDao.insertActiveTasks(this)
+            }
         return ids
     }
 
     override suspend fun deleteActiveTasks(task: List<Task>): Int =
-        activeTaskDao.deleteActiveTasks(
-            task
-                .filter { it.activeStatus != null }
-                .map { it.activeStatus!!.asEntity(it.id).id }
-        )
+        task
+            .filter { it.activeStatuses.isNotEmpty() }
+            .flatMap { it.activeStatuses.map { s -> s.asEntity(it.id).id } }
+            .run {
+                if (isNotEmpty()) activeTaskDao.deleteActiveTasks(this) else 0
+            }
 
     override suspend fun setActiveTaskAsCompleted(task: Task): Long =
-        task.activeStatus?.let {
+        task.activeStatuses.firstOrNull()?.let {
             activeTaskDao.insertActiveTasks(
                 listOf(it.copy(isCompleted = true).asEntity(task.id))
             ).firstOrNull() ?: 0
